niri,update,pkgs,logitech,ghostty,hyprland,misc,justfile
- niri: config overhaul
  + move to nix-based configuration, rather than native kdl
    * uses personal flake forked off
      https://codeberg.org/BANanaD3V/niri-nix
    * should move to home-manager when merged upstream (needs KDL
      support in nixpkgs)
  + overlays moved to niri directory (from misc), added patches
    * maximize new windows by default, which allows switching to the
      default proportion of 0.5 with the maximize-column keybind
    * add several actions for better control of column operations
  + several keybind changes now that I'm comfortable with niri and know
    what I want (and have patches to add that)
  + minor improvements to startup.sh
- update: pkgs
  + pkgs: exiled-exchange-2
- pkgs: add vesktop, zulip, element-desktop (user)
  + switch to using these from nixpkgs rather than flatpak
- logitech: un-invert thumbwheel so down=right
- ghostty: remove unused overlay patches
- hyprland: remove unused overlay patch (for grimblast)
- misc: remove unnecessary "builtin."
- justfile: don't use nom when activating with nixos-rebuild
  + nom only really matters when there's stuff to be built, but that
    already happened with the build job
  + permits color highlighting in output

+++ home/niri/startup.sh
@@ -5,16 +5,18 @@ trap 'pkill -eu mithic -KILL; exit 2' INT
 trap '' HUP TERM QUIT CONT TSTP
 
 niri-session >>/tmp/niri.log 2>&1
-
 sleep 1
-systemctl --user exit
 
-sleep 5
+systemctl --user exit
+sleep 1
 pgrep -u mithic --list-name --ignore-ancestors
+[ $? -eq 1 ] && exit 0
+sleep 1
 pstree mithic
-sleep 5
+sleep 10
+
 pkill -eu mithic -TERM --ignore-ancestors
-timeout '10s' pidwait -u mithic --ignore-ancestors
+timeout -f '10s' pidwait -u mithic --ignore-ancestors
 [ $? -le 1 ] && exit 0
 
 pgrep -u mithic --list-name --ignore-ancestors
